Otimização de componentes;
Criados components: EmailInput, FullNameInput, PasswordInput;

Os componentes: EmailInput, FullNameInput, PasswordInput e MaskedInput, passaram a utilizar os estilos padrão do componente TextInput;

Substituições dos componentes nas páginas: SignIn, SignUp e PasswordRecovery;
