+++ src/components/inputs/masked-input.tsx
@@ -1,0 +1,10 @@
+import { TextInputMask, TextInputMaskProps } from "react-native-masked-text"
+import { TextInputStyles as styles } from './text-input'
+
+interface Properties extends TextInputMaskProps { }
+
+export function MaskedInput(properties: Properties) {
+    return (
+        <TextInputMask {...properties} style={styles.input} />
+    )
+}
