+++ src/pages/SignIn/index.tsx
@@ -1,13 +1,13 @@
 import { useContext, useState } from "react";
 import { StyleSheet, SafeAreaView, View, Image, Text } from "react-native";
-
 import { StackScreenProps } from "@react-navigation/stack";
 import { AuthContext } from "../../contexts/auth.provider";
 import { StackParams } from "../../types/stack.params";
-import { TextInput } from "../../components/inputs/text-input";
 import { Button } from "../../components/buttons/button";
 import { StatusBar } from "expo-status-bar";
 import { CommandLink } from "../../components/buttons/command-link";
+import { EmailInput } from "../../components/inputs/email-input";
+import { PasswordInput } from "../../components/inputs/password-input";
 
 interface Properties extends StackScreenProps<StackParams, "SignIn"> { }
 
@@ -31,23 +31,12 @@ export default function SignIn({ navigation }: Properties) {
           <Text style={styles.title}>seja bem-vindo(a).</Text>
         </View>
         <View style={{ marginBottom: 40 }}>
-          <TextInput
+          <EmailInput
             value={email}
-            onChangeText={setEmail}
-            textContentType="emailAddress"
-            keyboardType="email-address"
-            autoCapitalize="none"
-            placeholder="E-mail"
-            placeholderTextColor="#1F537E"
-            maxFontSizeMultiplier={14} />
-          <TextInput
+            onChangeText={setEmail} />
+          <PasswordInput
             value={password}
-            onChangeText={setPassword}
-            textContentType='password'
-            secureTextEntry
-            autoCapitalize="none"
-            placeholder='Senha'
-            placeholderTextColor='#1F537E' />
+            onChangeText={setPassword} />
           <Button onPress={handleSingIn} title='ENTRAR' />
         </View>
         <CommandLink
